+++ target_bigquery/bigquery_schema.py
@@ -27,34 +27,6 @@ METADATA_FIELDS = {
 }
 
 
-def cleanup_record(schema, record):
-    """
-    Clean up / prettify field names, make sure they match BigQuery naming conventions.
-
-    :param JSON record generated by the tap and piped into target-bigquery
-    :param bq_schema: JSON schema generated by the tap and piped into target-bigquery
-    :return: JSON record/data, where field names are cleaned up / prettified.
-    """
-    if not isinstance(record, dict) and not isinstance(record, list):
-        return record
-
-    elif isinstance(record, list):
-        nr = []
-        for item in record:
-            nr.append(cleanup_record(schema, item))
-        return nr
-
-    elif isinstance(record, dict):
-        nr = {}
-        for key, value in record.items():
-            nkey = create_valid_bigquery_name(key)
-            nr[nkey] = cleanup_record(schema, value)
-        return nr
-
-    else:
-        raise Exception(f"unhandled instance of record: {record}")
-
-
 def create_valid_bigquery_name(column_name: str) -> str:
     """
     Transforms a given column name into a BigQuery-compliant column name.
