+++ target_bigquery/process.py
@@ -3,14 +3,16 @@ import singer
 import pyarrow as pa
 import pyarrow.parquet as pq
 from typing import Any, Dict, List
-from datetime import datetime, UTC
-from dateutil import parser as dateutil_parser
+from datetime import datetime
 
 from target_bigquery.config import TargetConfig, TablesConfig, TableConfig
 from target_bigquery.bigquery_schema import (
     build_schema,
     create_valid_bigquery_name,
-    cleanup_record,
+)
+from target_bigquery.pyarrow_schema import (
+    convert_and_filter_record_to_pyarrow,
+    bq_schema_to_pyarrow_schema,
 )
 from target_bigquery.simplify_json_schema import simplify
 from target_bigquery.validate_json_schema import (
@@ -26,274 +28,43 @@ logger = singer.get_logger()
 STREAM_ROW_CACHE_SIZE = 75000
 
 
-def convert_value_to_pyarrow_type(value: Any, expected_type: pa.DataType) -> Any:
+def cleanup_record(schema, record):
     """
-    Convert a value to match the expected PyArrow type.
-    Handle common type conversion issues gracefully.
-    Incorporates best practices from target-parquet implementation.
+    Recursively sanitize field names in a record to comply with BigQuery naming conventions.
+
+    This function processes incoming Singer tap records by transforming all field names
+    to meet BigQuery column naming requirements. It handles nested objects and arrays
+    recursively, ensuring all field names are valid BigQuery identifiers.
+
+    BigQuery naming rules applied:
+    - Replace non-alphanumeric characters (except underscores) with underscores
+    - Ensure field names start with letter or underscore
+    - Truncate to 300 characters maximum
+    - Avoid reserved prefixes
+
+    :param schema: JSON schema for the record (currently unused but kept for API compatibility)
+    :param record: Raw record data from Singer tap (dict, list, or primitive value)
+    :return: Record with sanitized field names that comply with BigQuery naming conventions
+    :raises Exception: If record contains unsupported data types
     """
-    if value is None:
-        return None
+    if not isinstance(record, dict) and not isinstance(record, list):
+        return record
 
-    try:
-        # Handle empty strings for non-string types (key insight from target-parquet)
-        if (
-            isinstance(value, str)
-            and value.strip() == ""
-            and not pa.types.is_string(expected_type)
-        ):
-            return None
+    elif isinstance(record, list):
+        nr = []
+        for item in record:
+            nr.append(cleanup_record(schema, item))
+        return nr
 
-        # Handle timestamp conversion using more robust dateutil parser
-        if pa.types.is_timestamp(expected_type):
-            if isinstance(value, str):
-                if value.lower() == "null":
-                    return None
-                try:
-                    # Use dateutil parser which is much more flexible than fromisoformat
-                    return dateutil_parser.parse(value)
-                except (ValueError, TypeError) as e:
-                    logger.warning(
-                        f"Cannot convert string '{value}' to timestamp: {e}, setting to None"
-                    )
-                    return None
-            elif isinstance(value, datetime):
-                return value
+    elif isinstance(record, dict):
+        nr = {}
+        for key, value in record.items():
+            nkey = create_valid_bigquery_name(key)
+            nr[nkey] = cleanup_record(schema, value)
+        return nr
 
-        # Handle integer conversion
-        elif pa.types.is_integer(expected_type):
-            if isinstance(value, str):
-                if value.lower() == "null":
-                    return None
-                try:
-                    # Handle potential float strings like "123.0"
-                    return int(float(value))
-                except (ValueError, TypeError):
-                    logger.warning(
-                        f"Cannot convert string '{value}' to integer, setting to None"
-                    )
-                    return None
-            elif isinstance(value, (int, float)):
-                return int(value)
-
-        # Handle float conversion
-        elif pa.types.is_floating(expected_type):
-            if isinstance(value, str):
-                if value.lower() == "null":
-                    return None
-                try:
-                    return float(value)
-                except (ValueError, TypeError):
-                    logger.warning(
-                        f"Cannot convert string '{value}' to float, setting to None"
-                    )
-                    return None
-            elif isinstance(value, (int, float)):
-                return float(value)
-            elif hasattr(value, "__class__") and value.__class__.__name__ == "Decimal":
-                try:
-                    return float(value)
-                except (ValueError, TypeError):
-                    logger.warning(
-                        f"Cannot convert Decimal '{value}' to float, setting to None"
-                    )
-                    return None
-
-        # Handle boolean conversion
-        elif pa.types.is_boolean(expected_type):
-            if isinstance(value, str):
-                if value.lower() == "null":
-                    return None
-                value_lower = value.lower()
-                if value_lower in ("true", "1", "yes", "y", "t"):
-                    return True
-                elif value_lower in ("false", "0", "no", "n", "f"):
-                    return False
-                else:
-                    logger.warning(
-                        f"Cannot convert string '{value}' to boolean, setting to None"
-                    )
-                    return None
-            elif isinstance(value, bool):
-                return value
-
-        # Handle string conversion (most permissive)
-        elif pa.types.is_string(expected_type):
-            if value is None:
-                return None
-            # Handle complex objects by JSON serialization (from target-parquet)
-            if isinstance(value, (list, dict)):
-                try:
-                    return json.dumps(value, default=str)
-                except (TypeError, ValueError):
-                    return str(value)
-            return str(value)
-
-        # For other types, return as-is and let PyArrow handle it
-        return value
-
-    except Exception as e:
-        logger.warning(f"Error converting value '{value}' to type {expected_type}: {e}")
-        return None
-
-
-def convert_and_filter_record_to_pyarrow(
-    record: Any, schema: pa.Schema, bq_schema_dict: Dict[str, Any] = None
-) -> Dict[str, Any]:
-    """
-    Unified function that combines schema filtering and PyArrow type conversion.
-    Handles nested records and arrays recursively.
-
-    :param record: Input record (dict, list, or primitive value)
-    :param schema: PyArrow Schema defining expected structure
-    :param bq_schema_dict: BigQuery schema dict for field filtering (optional)
-    :return: Converted record matching PyArrow schema
-    """
-
-    # Handle non-dict records (for recursive calls with lists/primitives)
-    if not isinstance(record, dict):
-        if isinstance(record, list):
-            # Handle array case - should have list type in schema
-            # For now, return as-is and let PyArrow handle it
-            return record
-        else:
-            # Primitive value - should not happen at top level
-            return record
-
-    converted_row = {}
-
-    # Process each field in the PyArrow schema
-    for field in schema:
-        field_name = field.name
-        value = record.get(field_name)
-
-        # If field is not in record, set to None as default
-        if field_name not in record:
-            converted_row[field_name] = None
-            continue
-
-        if value is None:
-            converted_row[field_name] = None
-            continue
-
-        # Handle different PyArrow field types
-        if pa.types.is_struct(field.type):
-            # Nested record - recurse
-            if isinstance(value, dict):
-                nested_schema = pa.schema(field.type)
-                nested_bq_schema = (
-                    bq_schema_dict.get(field_name, {}).get("fields", {})
-                    if bq_schema_dict
-                    else {}
-                )
-                converted_row[field_name] = convert_and_filter_record_to_pyarrow(
-                    value, nested_schema, nested_bq_schema
-                )
-            else:
-                logger.warning(
-                    f"Expected dict for struct field {field_name}, got {type(value)}"
-                )
-                converted_row[field_name] = None
-
-        elif pa.types.is_list(field.type):
-            # Array field - convert each element
-            if isinstance(value, list):
-                element_type = field.type.value_type
-                converted_elements = []
-
-                for element in value:
-                    if pa.types.is_struct(element_type):
-                        # Array of nested records
-                        nested_schema = pa.schema(element_type)
-                        nested_bq_schema = (
-                            bq_schema_dict.get(field_name, {}).get("fields", {})
-                            if bq_schema_dict
-                            else {}
-                        )
-                        converted_element = convert_and_filter_record_to_pyarrow(
-                            element, nested_schema, nested_bq_schema
-                        )
-                    else:
-                        # Array of primitives
-                        converted_element = convert_value_to_pyarrow_type(
-                            element, element_type
-                        )
-                    converted_elements.append(converted_element)
-
-                converted_row[field_name] = converted_elements
-            else:
-                logger.warning(
-                    f"Expected list for array field {field_name}, got {type(value)}"
-                )
-                converted_row[field_name] = None
-
-        else:
-            # Simple field - use existing conversion logic
-            converted_value = convert_value_to_pyarrow_type(value, field.type)
-            converted_row[field_name] = converted_value
-
-    return converted_row
-
-
-def bq_field_to_pyarrow_field(bq_field: SchemaField) -> pa.Field:
-    """
-    Convert a BigQuery SchemaField to a PyArrow Field
-
-    :param bq_field: BigQuery SchemaField object
-    :return: PyArrow Field
-    """
-    field_name = bq_field.name
-    field_type = bq_field.field_type
-    field_mode = bq_field.mode
-    precision = bq_field.precision
-    scale = bq_field.scale
-
-    # Type mapping from BigQuery to PyArrow
-    type_mapping = {
-        "STRING": pa.string(),
-        "INTEGER": pa.int64(),
-        "FLOAT": pa.float64(),
-        "BOOLEAN": pa.bool_(),
-        "TIMESTAMP": pa.timestamp("us", tz="UTC"),
-        "DATE": pa.date32(),
-        "TIME": pa.time64("us"),
-        "GEOGRAPHY": pa.string(),
-        "DECIMAL": pa.decimal128(precision or 38, scale or 9),
-        "BIGDECIMAL": pa.decimal256(precision or 76, scale or 38),
-    }
-
-    if field_type == "RECORD":
-        # Handle nested records recursively
-        nested_fields = []
-        for nested_field in bq_field.fields:
-            nested_fields.append(bq_field_to_pyarrow_field(nested_field))
-
-        arrow_type = pa.struct(nested_fields)
     else:
-        arrow_type = type_mapping.get(field_type, pa.string())
-
-    # Handle REPEATED mode
-    if field_mode == "REPEATED":
-        arrow_type = pa.list_(arrow_type)
-
-    # Determine nullability
-    nullable = field_mode == "NULLABLE"
-
-    return pa.field(field_name, arrow_type, nullable=nullable)
-
-
-def bq_schema_to_pyarrow_schema(bq_fields: List[SchemaField]) -> pa.Schema:
-    """
-    Convert BigQuery schema (list of SchemaFields) to PyArrow schema
-
-    :param bq_fields: List of BigQuery SchemaField objects
-    :return: PyArrow Schema
-    """
-    pa_fields = []
-    for bq_field in bq_fields:
-        pa_fields.append(bq_field_to_pyarrow_field(bq_field))
-
-    return pa.schema(pa_fields)
+        raise Exception(f"unhandled instance of record: {record}")
 
 
 def build_table_name(
@@ -318,6 +89,9 @@ def build_bq_schema_dict(schema):
     schema_dict = {}
     for field in schema:
         f = field if isinstance(field, dict) else field.to_api_repr()
+        # Make a copy to avoid mutating the original field
+        import copy
+        f = copy.deepcopy(f)
         schema_dict[f["name"]] = f
         if f.get("fields"):
             schema_dict[f["name"]]["fields"] = build_bq_schema_dict(f["fields"])
@@ -368,50 +142,35 @@ class SingerProcessor:
             return
 
         try:
-            # Convert all rows to match PyArrow types
-            try:
-                converted_rows = [
-                    convert_and_filter_record_to_pyarrow(
-                        row,
-                        self.pyarrow_schemas[stream_name],
-                        self.big_query_schema_dicts[stream_name],
-                    )
-                    for row in self.rows[stream_name]
-                ]
+            table = pa.Table.from_pylist(
+                self.rows[stream_name], schema=self.pyarrow_schemas[stream_name]
+            )
 
-                # Create PyArrow table from converted rows
-                table = pa.Table.from_pylist(
-                    converted_rows, schema=self.pyarrow_schemas[stream_name]
-                )
+        except Exception as conversion_error:
+            # Log detailed information about the data that's causing issues
+            logger.error(
+                f"Error converting or creating PyArrow table for stream {stream_name}: {conversion_error}"
+            )
 
-            except Exception as conversion_error:
-                # Log detailed information about the data that's causing issues
+            # Try to identify which column is causing the issue
+            schema_fields = self.pyarrow_schemas[stream_name]
+            sample_row = self.rows[stream_name][0] if self.rows[stream_name] else {}
+
+            for field in schema_fields:
+                field_name = field.name
+                field_type = field.type
+                sample_value = sample_row.get(field_name)
+
                 logger.error(
-                    f"Error converting or creating PyArrow table for stream {stream_name}: {conversion_error}"
+                    f"Field: {field_name}, Expected type: {field_type}, Sample value: {sample_value} (type: {type(sample_value)})"
                 )
 
-                # Try to identify which column is causing the issue
-                schema_fields = self.pyarrow_schemas[stream_name]
-                sample_row = self.rows[stream_name][0] if self.rows[stream_name] else {}
+            raise conversion_error
 
-                for field in schema_fields:
-                    field_name = field.name
-                    field_type = field.type
-                    sample_value = sample_row.get(field_name)
-
-                    logger.error(
-                        f"Field: {field_name}, Expected type: {field_type}, Sample value: {sample_value} (type: {type(sample_value)})"
-                    )
-
-                raise conversion_error
-
+        try:
             # Get writer and write table
             writer = self._get_parquet_writer(stream_name)
             writer.write_table(table)
-
-            # Close writer immediately to free memory and file handles
-            writer.close()
-            self.parquet_writers.pop(stream_name, None)
 
             # Clear the buffer
             row_count = len(self.rows[stream_name])
@@ -526,7 +285,13 @@ class SingerProcessor:
             )
             nr["_time_loaded"] = datetime.utcnow().isoformat()
 
-        self.rows[stream_name].append(nr)
+        converted_row = convert_and_filter_record_to_pyarrow(
+            nr,
+            self.pyarrow_schemas[stream_name],
+            self.big_query_schema_dicts[stream_name],
+        )
+
+        self.rows[stream_name].append(converted_row)
         self.total_cached_rows += 1
 
         # Flush all streams when total cached rows exceed threshold
